Update README.md
What I like about Dont mess with the zohan is that Zohan Dvir is a promiscuous, superhuman Israeli counter-terrorist of the Israel Defense Forces. Despite his success and popularity, he has grown tired of the everlasting conflicts in his country and dreams of becoming a hairstylist at Paul Mitchell's in the United States. He finally admits to his parents his desire to cut hair, but they make fun of him.